Changed grade float to decimal on upload page

+++ CodeForACause/StlCollegePrepWebsite/Models/CourseStudent.cs
@@ -17,7 +17,7 @@ namespace StlCollegePrepWebsite.Models
         public string StudentId { get; set; }
 
         public int? AcademicYear { get; set; }
-        public double? AwardedGrade { get; set; }
+        public decimal? AwardedGrade { get; set; }
 
         // Navigation properties
         [ForeignKey("CourseId")]
